导出Excel报表接口业务代码开发

+++ sky-server/src/main/java/com/sky/mapper/UserMapper.java
@@ -15,5 +15,5 @@ public interface UserMapper {
 
     void insert(User user);
 
-    Double userStatistics(Map<String, Object> map);
+    Integer userStatistics(Map<String, Object> map);
 }
